backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-cors==0.8.0
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bcrypt==4.0.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
frozenlist==1.5.0
grpcio==1.59.3
grpcio-tools==1.59.3
idna==3.10
iniconfig==2.1.0
motor==3.3.1
multidict==6.2.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.3.1
protobuf==4.25.6
pyasn1==0.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.6.1
pytest==8.3.5
pytest-aiohttp==1.1.0
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==78.1.0
six==1.17.0
typing_extensions==4.13.0
websockets==12.0
yarl==1.18.3
